- Execute requirements with atomic start/tick/end transactions (commit/rollback semantics) - Add per-process requirement overlay and overlay modifier pipeline (incl. NBT roundtrip) - Support recipe groups lookup and processor scanning wiring - Add unit tests for atomicity, per-tick energy/fluid IO, overlays, and group queries - Add Structure Projection Preview documentation (/pm_preview, 24-way orientation, modes) - Update machine/structure docs for area-loaded fast-fail and transactional requirements/overlay - Fix UI doc lang filename casing; add keybinding translations (en_us/zh_cn)

+++ src/main/kotlin/api/recipe/process/component/type/ParallelProcessComponent.kt
@@ -4,6 +4,6 @@ import github.kasuminova.prototypemachinery.api.recipe.process.component.RecipeP
 
 public interface ParallelProcessComponent : RecipeProcessComponent {
 
-    public val parallelism: Int
+    public val parallelism: Long
 
 }
